+++ AMI/get_latest_amis.py
@@ -1,11 +1,13 @@
 """
-Find AMI ID of the latest
-- ECS optimized Bottlerocket AMIs
-- EKS optimized Bottlerocket AMIs
-- EC2 Linux AMIs
-- ECS optimized Amazon Linux 1/2 AMIs
-- EKS optimized Amazon Linux 2 AMIs
-- EC2 Windows AMIs
+Find AMI ID of the latest AMI
+- EC2 Amazon Linux AMIs
+- EC2 Windows Server AMIs
+- ECS-optimized Amazon Linux AMIs
+- ECS-optimized Bottlerocket AMIs
+- ECS-optimized Windows Server AMIs
+- EKS-optimized Amazon Linux AMIs
+- EKS-optimized Bottlerocket AMIs
+- EKS-optimized Windows Server AMIs
 """
 import json
 
@@ -13,34 +15,44 @@ import click
 from boto3.session import Session
 from helper.selector import prompt_multi_selection
 
-
-def get_bottlerocket_ecs_meta_dict():
-    # https://docs.aws.amazon.com/AmazonECS/latest/developerguide/ecs-bottlerocket.html
-    return {
-        "/aws/service/bottlerocket/aws-ecs-1/arm64": "Amazon ECS-Optimized Bottlerocket (arm64) AMI",
-        "/aws/service/bottlerocket/aws-ecs-1/x86_64": "Amazon ECS-Optimized Bottlerocket (x86_64) AMI",
-        "/aws/service/bottlerocket/aws-ecs-1-nvidia/arm64": "Amazon ECS-Optimized (NVIDIA GPU support) Bottlerocket (arm64) AMI",
-        "/aws/service/bottlerocket/aws-ecs-1-nvidia/x86_64": "Amazon ECS-Optimized (NVIDIA GPU support) Bottlerocket (x86_64) AMI",
-    }
+TOPIC_A = "AmazonLinux"
+TOPIC_B = "Bottlerocket"
+TOPIC_W = "Windows"
 
 
-def get_ecs_meta_dict():
-    # https://docs.aws.amazon.com/AmazonECS/latest/developerguide/retrieve-ecs-optimized_AMI.html
-    return {
-        "/aws/service/ecs/optimized-ami/amazon-linux": "Amazon ECS-optimized Amazon Linux (x86_64) AMI deprecated as of 2021-04-15",
-        "/aws/service/ecs/optimized-ami/amazon-linux-2": "Amazon ECS-Optimized Amazon Linux 2 (x86_64) AMI",
-        "/aws/service/ecs/optimized-ami/amazon-linux-2/arm64": "Amazon ECS-Optimized Amazon Linux 2 (arm64) AMI",
-        "/aws/service/ecs/optimized-ami/amazon-linux-2/gpu": "Amazon ECS-Optimized Amazon Linux 2 (GPU) AMI",
-        "/aws/service/ecs/optimized-ami/amazon-linux-2/inf": "Amazon ECS-Optimized Amazon Linux 2 (Inferentia) AMI",
-        "/aws/service/ecs/optimized-ami/amazon-linux-2022": "Amazon ECS-Optimized Amazon Linux 2022 (x86_64) AMI"
-    }
+def get_ecs_meta_dict(topic=TOPIC_A):
+    if topic == TOPIC_A:
+        # https://docs.aws.amazon.com/AmazonECS/latest/developerguide/retrieve-ecs-optimized_AMI.html
+        return {
+            "/aws/service/ecs/optimized-ami/amazon-linux-2022/recommended": "Amazon ECS-Optimized Amazon Linux 2022 (x86_64) AMI",
+            "/aws/service/ecs/optimized-ami/amazon-linux-2/recommended": "Amazon ECS-Optimized Amazon Linux 2 (x86_64) AMI",
+            "/aws/service/ecs/optimized-ami/amazon-linux-2/arm64/recommended": "Amazon ECS-Optimized Amazon Linux 2 (arm64) AMI",
+            "/aws/service/ecs/optimized-ami/amazon-linux-2/gpu/recommended": "Amazon ECS-Optimized Amazon Linux 2 (GPU) AMI",
+            "/aws/service/ecs/optimized-ami/amazon-linux-2/inf/recommended": "Amazon ECS-Optimized Amazon Linux 2 (Inferentia) AMI",
+            "/aws/service/ecs/optimized-ami/amazon-linux/recommended": "Amazon ECS-optimized Amazon Linux (x86_64) AMI deprecated as of 2021-04-15",
+        }
+    if topic == TOPIC_B:
+        # https://docs.aws.amazon.com/AmazonECS/latest/developerguide/ecs-bottlerocket.html
+        return {
+            "/aws/service/bottlerocket/aws-ecs-1/arm64/latest": "Amazon ECS-Optimized Bottlerocket (arm64) AMI",
+            "/aws/service/bottlerocket/aws-ecs-1/x86_64/latest": "Amazon ECS-Optimized Bottlerocket (x86_64) AMI",
+            "/aws/service/bottlerocket/aws-ecs-1-nvidia/arm64/latest": "Amazon ECS-Optimized (NVIDIA GPU support) Bottlerocket (arm64) AMI",
+            "/aws/service/bottlerocket/aws-ecs-1-nvidia/x86_64/latest": "Amazon ECS-Optimized (NVIDIA GPU support) Bottlerocket (x86_64) AMI",
+        }
+    if topic == TOPIC_W:
+        # https://docs.aws.amazon.com/AmazonECS/latest/developerguide/ecs-windows-ami-versions.html
+        return {
+            "/aws/service/ami-windows-latest/Windows_Server-2022-English-Core-ECS_Optimized": "Amazon ECS-optimized Windows Server 2022 Core AMI",
+            "/aws/service/ami-windows-latest/Windows_Server-2022-English-Full-ECS_Optimized": "Amazon ECS-optimized Windows Server 2022 Full AMI",
+            "/aws/service/ami-windows-latest/Windows_Server-2019-English-Core-ECS_Optimized": "Amazon ECS-optimized Windows Server 2019 Core AMI",
+            "/aws/service/ami-windows-latest/Windows_Server-2019-English-Full-ECS_Optimized": "Amazon ECS-optimized Windows Server 2019 Full AMI",
+            "/aws/service/ami-windows-latest/Windows_Server-2016-English-Full-ECS_Optimized": "Amazon ECS-optimized Windows Server 2016 Full AMI",
+        }
 
-
-def get_eks_meta_dict(get_bottlerocket=False):
-    # https://docs.aws.amazon.com/eks/latest/userguide/eks-linux-ami-versions.html
+def get_eks_meta_dict(topic=TOPIC_A):
     K8S_VERSIONS = [
         "1.23",
-        "1.22",K8S_VERSIONS
+        "1.22",
         "1.21",
         "1.20",
         "1.19",
@@ -48,19 +60,25 @@ def get_eks_meta_dict(get_bottlerocket=False):
         "1.17",
     ]
 
-    # https://docs.aws.amazon.com/eks/latest/userguide/retrieve-ami-id.html
     ami_variants = {}
-    if get_bottlerocket:
+    if topic == TOPIC_A:
+        # https://docs.aws.amazon.com/eks/latest/userguide/eks-linux-ami-versions.html
         for k8s_version in K8S_VERSIONS:
-            ami_variants[f"/aws/service/bottlerocket/aws-k8s-{k8s_version}/arm64"] = "Amazon EKS-optimized Bottlerocket (arm64 Standard) AMI"
-            ami_variants[f"/aws/service/bottlerocket/aws-k8s-{k8s_version}/x86_64"] = "Amazon EKS-optimized Bottlerocket (x86_64 Standard) AMI"
-            ami_variants[f"/aws/service/bottlerocket/aws-k8s-{k8s_version}-nvidia/arm64"] = "Amazon EKS-optimized Bottlerocket (arm64 NVIDIA) AMI"
-            ami_variants[f"/aws/service/bottlerocket/aws-k8s-{k8s_version}-nvidia/x86_64"] = "Amazon EKS-optimized Bottlerocket (x86_64 NVIDIA) AMI"
-    else:
+            ami_variants[f"/aws/service/eks/optimized-ami/{k8s_version}/amazon-linux-2/recommended"] = "Amazon EKS-optimized Amazon Linux 2 (x86_64) AMI"
+            ami_variants[f"/aws/service/eks/optimized-ami/{k8s_version}/amazon-linux-2-arm64/recommended"] = "Amazon EKS-optimized Amazon Linux 2 (arm64) AMI"
+            ami_variants[f"/aws/service/eks/optimized-ami/{k8s_version}/amazon-linux-2-gpu/recommended"] = "Amazon EKS-optimized Amazon Linux 2 (GPU) AMI"
+    elif topic == TOPIC_B:
+        # https://docs.aws.amazon.com/eks/latest/userguide/retrieve-ami-id-bottlerocket.html
         for k8s_version in K8S_VERSIONS:
-            ami_variants[f"/aws/service/eks/optimized-ami/{k8s_version}/amazon-linux-2"] = "Amazon EKS-optimized Amazon Linux 2 (x86_64) AMI"
-            ami_variants[f"/aws/service/eks/optimized-ami/{k8s_version}/amazon-linux-2-arm64"] = "Amazon EKS-optimized Amazon Linux 2 (arm64) AMI"
-            ami_variants[f"/aws/service/eks/optimized-ami/{k8s_version}/amazon-linux-2-gpu"] = "Amazon EKS-optimized Amazon Linux 2 (GPU) AMI"
+            ami_variants[f"/aws/service/bottlerocket/aws-k8s-{k8s_version}/arm64/latest"] = "Amazon EKS-optimized Bottlerocket (arm64 Standard) AMI"
+            ami_variants[f"/aws/service/bottlerocket/aws-k8s-{k8s_version}/x86_64/latest"] = "Amazon EKS-optimized Bottlerocket (x86_64 Standard) AMI"
+            ami_variants[f"/aws/service/bottlerocket/aws-k8s-{k8s_version}-nvidia/arm64/latest"] = "Amazon EKS-optimized Bottlerocket (arm64 NVIDIA) AMI"
+            ami_variants[f"/aws/service/bottlerocket/aws-k8s-{k8s_version}-nvidia/x86_64/latest"] = "Amazon EKS-optimized Bottlerocket (x86_64 NVIDIA) AMI"
+    elif topic == TOPIC_W:
+        # https://docs.aws.amazon.com/eks/latest/userguide/eks-ami-versions-windows.html
+        for k8s_version in K8S_VERSIONS:
+            ami_variants[f"/aws/service/ami-windows-latest/Windows_Server-2019-English-Core-EKS_Optimized-{k8s_version}"] = "Amazon EKS-optimized Windows Server 2019 Core AMI"
+            ami_variants[f"/aws/service/ami-windows-latest/Windows_Server-2019-English-Full-EKS_Optimized-{k8s_version}"] = "Amazon EKS-optimized Windows Server 2019 Full AMI"
 
     return ami_variants
 
@@ -94,6 +112,20 @@ def get_parameters(param_path, region, session):
             print(f"Skip region {region} due to error: {e}")
 
 
+def get_amis_from_get_parameters_by_path(param_path, region, session):
+    data = get_parameters_by_path(param_path=param_path, region=region, session=session)
+    if data:
+        resp = prompt_multi_selection("AMI", options=list(data.keys()), pre_selected_options=[])
+        for arn in resp.get("AMIs", []):
+            print(json.dumps(data[arn], default=str, indent=2, sort_keys=True))
+
+
+def get_amis_from_get_parameters(ami_dict, region, session):
+    resp = prompt_multi_selection("AMI", options=list(ami_dict.keys()), pre_selected_options=[])
+    for name in resp.get("AMIs", []):
+        get_parameters(param_path=f"{name}/image_id", region=region, session=session)
+
+
 @click.group(help="List the lastest AWS managed AMIs")
 @click.option("--profile", "-p", default="default", show_default=True, help="AWS profile name")
 @click.option("--region", "-r", default="ap-southeast-2", show_default=True, help="AWS region, or 'all' for all regions")
@@ -103,62 +135,54 @@ def cli_main(ctx, profile, region):
     ctx.obj = {"session": session, "region": region}
 
 
-@cli_main.command(help="Find ECS optimized Bottlerocket AMIs")
+@cli_main.command(help="Find Amazon EC2 Amazon Linux AMIs")
 @click.pass_context
-def bottlerocket_ecs(ctx):
-    region, session = ctx.obj["region"], ctx.obj["session"]
-    resp = prompt_multi_selection("AMI", options=list(get_bottlerocket_ecs_meta_dict().keys()), pre_selected_options=[])
-    for name in resp.get("AMIs", []):
-        get_parameters(param_path=f"{name}/latest/image_id", region=region, session=session)
+def ec2_amazon_linux(ctx):
+    # https://docs.aws.amazon.com/AmazonECS/latest/developerguide/ecs-ami-versions.html
+    get_amis_from_get_parameters_by_path("/aws/service/ami-amazon-linux-latest", ctx.obj["region"], ctx.obj["session"])
 
 
-@cli_main.command(help="Find EKS optimized Bottlerocket AMIs")
+@cli_main.command(help="Find Amazon EC2 Windows Server AMIs")
 @click.pass_context
-def bottlerocket_eks(ctx):
-    region, session = ctx.obj["region"], ctx.obj["session"]
-    resp = prompt_multi_selection("AMI", options=list(get_eks_meta_dict(True).keys()), pre_selected_options=[])
-    for name in resp.get("AMIs", []):
-        get_parameters(param_path=f"{name}/latest/image_id", region=region, session=session)
+def ec2_windows(ctx):
+    # https://docs.aws.amazon.com/AWSEC2/latest/WindowsGuide/finding-an-ami.html#finding-an-ami-aws-cli
+    get_amis_from_get_parameters_by_path("/aws/service/ami-windows-latest", ctx.obj["region"], ctx.obj["session"])
 
 
-@cli_main.command(help="Find EC2 Linux AMIs")
+@cli_main.command(help="Find Amazon ECS-optimized Amazon Linux AMIs")
 @click.pass_context
-def ec2(ctx):
-    region, session = ctx.obj["region"], ctx.obj["session"]
-    data = get_parameters_by_path(param_path="/aws/service/ami-amazon-linux-latest", region=region, session=session)
-    if data:
-        resp = prompt_multi_selection("AMI", options=list(data.keys()), pre_selected_options=[])
-        for arn in resp.get("AMIs", []):
-            print(json.dumps(data[arn], default=str, indent=2, sort_keys=True))
+def ecs_amazon_linux(ctx):
+    get_amis_from_get_parameters(get_ecs_meta_dict(TOPIC_A), ctx.obj["region"], ctx.obj["session"])
 
 
-@cli_main.command(help="Find ECS optimized AMIs")
+@cli_main.command(help="Find Amazon ECS-optimized Bottlerocket AMIs")
 @click.pass_context
-def ecs(ctx):
-    region, session = ctx.obj["region"], ctx.obj["session"]
-    resp = prompt_multi_selection("AMI", options=list(get_ecs_meta_dict().keys()), pre_selected_options=[])
-    for name in resp.get("AMIs", []):
-        get_parameters(param_path=f"{name}/recommended", region=region, session=session)
+def ecs_bottlerocket(ctx):
+    get_amis_from_get_parameters(get_ecs_meta_dict(TOPIC_B), ctx.obj["region"], ctx.obj["session"])
 
 
-@cli_main.command(help="Find EKS optimized AMIs")
+@cli_main.command(help="Find Amazon ECS-optimized Windows Server AMIs")
 @click.pass_context
-def eks(ctx):
-    region, session = ctx.obj["region"], ctx.obj["session"]
-    resp = prompt_multi_selection("AMI", options=list(get_eks_meta_dict().keys()), pre_selected_options=[])
-    for name in resp.get("AMIs", []):
-        get_parameters(param_path=f"{name}/recommended", region=region, session=session)
+def ecs_windows(ctx):
+    get_amis_from_get_parameters(get_ecs_meta_dict(TOPIC_W), ctx.obj["region"], ctx.obj["session"])
 
 
-@cli_main.command(help="Find EC2 Windows AMIs")
+@cli_main.command(help="Find Amazon EKS-optimized Amazon Linux AMIs")
 @click.pass_context
-def windows(ctx):
-    region, session = ctx.obj["region"], ctx.obj["session"]
-    data = get_parameters_by_path(param_path="/aws/service/ami-windows-latest", region=region, session=session)
-    if data:
-        resp = prompt_multi_selection("AMI", options=list(data.keys()), pre_selected_options=[])
-        for arn in resp.get("AMIs", []):
-            print(json.dumps(data[arn], default=str, indent=2, sort_keys=True))
+def eks_amazon_linux(ctx):
+    get_amis_from_get_parameters(get_eks_meta_dict(TOPIC_A), ctx.obj["region"], ctx.obj["session"])
+
+
+@cli_main.command(help="Find Amazon EKS-optimized Bottlerocket AMIs")
+@click.pass_context
+def eks_bottlerocket(ctx):
+    get_amis_from_get_parameters(get_eks_meta_dict(TOPIC_B), ctx.obj["region"], ctx.obj["session"])
+
+
+@cli_main.command(help="Find Amazon EKS-optimized Windows AMIs")
+@click.pass_context
+def eks_windows(ctx):
+    get_amis_from_get_parameters(get_eks_meta_dict(TOPIC_W), ctx.obj["region"], ctx.obj["session"])
 
 
 if __name__ == "__main__":